Fix #85 with non-empty redhat security source and empty-list checks.
The Red Hat security updates are deprecated, and the old source location now contains just an empty directory tree.  This caused Dagda to attempt to push empty lists into the MongoDB, which generated #85.  This fix both adds protections on the MongoDB write by not writing empty lists, and by changing the download URL to one that works.  Note that this is a temporary fix, because the Red Hat update indicates that this is still a deprecated source.  The v2 should be used instead.